+++ jvm/src/main/scala/gpdviz/async/NullPublisher.scala
@@ -1,9 +1,0 @@
-package gpdviz.async
-
-import gpdviz.Notif
-
-object NullPublisher extends Publisher {
-  def details: String = "no-publisher"
-
-  def publish(notif: Notif): Unit = ()
-}